`mapIt()` no longer calls `mapArray()`
