+++ src/transformation/index.ts
@@ -1,4 +1,5 @@
 export * from './filter-array';
 export * from './filter-it';
 export * from './flat-map-it';
+export * from './map-array';
 export * from './map-it';
