+++ src/transformation/map-it.spec.ts
@@ -1,4 +1,4 @@
-import { overMany, overNone } from '../construction';
+import { overMany } from '../construction';
 import { makePushIterator } from '../make-push-iterator';
 import { mapIt } from './map-it';
 
@@ -88,63 +88,6 @@ describe('mapIt', () => {
 
         expect([...mapIt(it, element => `${element}!`)]).toEqual(['test!']);
       });
-    });
-  });
-
-  describe('over array', () => {
-    it('converts elements', () => {
-      expect([...mapIt([11, 22, 33], element => `${element}!`)]).toEqual(['11!', '22!', '33!']);
-    });
-
-    describe('[Symbol.iterator]', () => {
-      it('converts elements', () => {
-
-        const it = mapIt([11, 22, 33], element => `${element}!`)[Symbol.iterator]();
-
-        expect([...it]).toEqual(['11!', '22!', '33!']);
-        expect(it[Symbol.iterator]()).toBe(it);
-      });
-    });
-
-    describe('forNext', () => {
-      it('reports converted elements', () => {
-
-        const result: string[] = [];
-
-        expect(mapIt([11, 22, 33], element => `${element}!`)[Symbol.iterator]().forNext(element => {
-          result.push(element);
-        })).toBe(false);
-        expect(result).toEqual(['11!', '22!', '33!']);
-      });
-      it('resumes conversion', () => {
-
-        const result: string[] = [];
-        const it = mapIt([11, 22, 33], element => `${element}!`)[Symbol.iterator]();
-
-        expect(it.forNext(() => false)).toBe(true);
-        expect(it.forNext(element => {
-          result.push(element);
-        })).toBe(false);
-        expect(result).toEqual(['22!', '33!']);
-
-        expect(it.forNext(element => {
-          result.push(element);
-        })).toBe(false);
-        expect(result).toEqual(['22!', '33!']);
-        expect([...it]).toHaveLength(0);
-      });
-    });
-  });
-
-  describe('over empty array', () => {
-    it('returns `overNone()`', () => {
-      expect(mapIt([], () => 'wrong')).toBe(overNone());
-    });
-  });
-
-  describe('over one-element array', () => {
-    it('iterates over single element', () => {
-      expect([...mapIt(['one'], el => el + '!')]).toEqual(['one!']);
     });
   });
 });
