+++ src/transformation/map-it.ts
@@ -2,7 +2,6 @@
  * @packageDocumentation
  * @module @proc7ts/push-iterator
  */
-import { overNone, overOne } from '../construction';
 import type { PushIterable, PushOrRawIterable } from '../push-iterable';
 import { isPushIterable, PushIterable__symbol } from '../push-iterable';
 import { PushIterator$iterator } from '../push-iterator.impl';
@@ -23,13 +22,7 @@ export function mapIt<T, R>(
     source: PushOrRawIterable<T>,
     convert: (this: void, element: T) => R,
 ): PushIterable<R> {
-  if (isPushIterable(source)) {
-    return mapPushIterable(source, convert);
-  }
-  if (Array.isArray(source)) {
-    return mapArray(source, convert);
-  }
-  return mapRawIterable(source, convert);
+  return isPushIterable(source) ? mapPushIterable(source, convert) : mapRawIterable(source, convert);
 }
 
 /**
@@ -68,65 +61,6 @@ function mapPushIterable<T, R>(source: PushIterable<T>, convert: (this: void, el
         },
 
         forNext: accept => it.forNext(element => accept(convert(element))),
-
-      };
-    },
-  };
-}
-
-/**
- * Creates a {@link PushIterable push iterable} with the results of calling a provided function on every element of the
- * given `array`.
- *
- * @typeParam T  A type of array elements.
- * @typeParam R  A type of converted elements.
- * @param array  A source array-like instance.
- * @param convert  A function that produces an element of new iterable, taking array element as the only parameter.
- *
- * @returns New push iterable of transformed elements.
- */
-export function mapArray<T, R>(
-    array: ArrayLike<T>,
-    convert: (this: void, element: T) => R,
-): PushIterable<R> {
-  if (array.length <= 1) {
-    if (!array.length) {
-      return overNone();
-    }
-    return overOne(convert(array[0]));
-  }
-
-  return {
-    [PushIterable__symbol]: 1,
-    [Symbol.iterator]() {
-
-      let i = 0;
-
-      return {
-
-        [PushIterable__symbol]: 1,
-
-        [Symbol.iterator]: PushIterator$iterator,
-
-        next: () => i < array.length ? { value: convert(array[i++]) } : { done: true } as IteratorReturnResult<R>,
-
-        forNext(accept) {
-          if (i >= array.length) {
-            return false;
-          }
-
-          for (; ;) {
-
-            const goOn = accept(convert(array[i++]));
-
-            if (i >= array.length) {
-              return false;
-            }
-            if (goOn === false) {
-              return true;
-            }
-          }
-        },
 
       };
     },
